LRSD-8348 [assets] Rename
