+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/assets/NavigationMenu/EnterpriseIconGray.tsx
