+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/assets/NavigationMenu/index.ts
@@ -1,0 +1,31 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export * from './AnalyticsIcon';
+export * from './AnalyticsIconGray';
+export * from '../AttachmentsIcon';
+export * from './BusinessEventsIcon';
+export * from './BusinessEventsIconGray'
+export * from './CommerceIcon';
+export * from './CommerceIconGray';
+export * from './DXPIcon';
+export * from './DXPIconGray';
+export * from './EnterpriseIcon';
+export * from './EnterpriseIconGray';
+export * from './ExperienceCloudIcon';
+export * from './ExperienceCloudIconGray';
+export * from './LXCIcon';
+export * from './LXCIconGray';
+export * from './OverviewIcon';
+export * from './OverviewIconGray';
+export * from './PartnershipIcon';
+export * from './PartnershipIconGray';
+export * from './PortalIcon';
+export * from './PortalIconGray';
+export * from './ProjectUsageIcon';
+export * from './ProjectUsageIconGray';
+export * from './ProductActivationIconGray';
+export * from './TeamMembersIcon';
+export * from './TeamMembersIconGray';
