+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/assets/index.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export * from './attachments_icon';
-export * from './icon_breadcrumb';
-export * from './intro_onboarding';
-export * from './navigation-menu';
-export * from './virtual_cluster';
+export * from './AttachmentsIcon';
+export * from './IconBreadcrumb';
+export * from './IntroOnboarding';
+export * from './NavigationMenu';
+export * from './VirtualCluster';
