--- workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/assets/navigation-menu/index.tsx
@@ -1,31 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-export * from './analytics_icon';
-export * from './analytics_icon_gray';
-export * from '../attachments_icon';
-export * from './business_events_icon';
-export * from './business_events_icon_gray'
-export * from './commerce_icon';
-export * from './commerce_icon_gray';
-export * from './dxp_icon';
-export * from './dxp_icon_gray';
-export * from './enterprise_icon';
-export * from './enterprise_icon_gray';
-export * from './experience_cloud_icon';
-export * from './experience_cloud_icon_gray';
-export * from './lxc_icon';
-export * from './lxc_icon_gray';
-export * from './overview_icon';
-export * from './overview_icon_gray';
-export * from './partnership_icon';
-export * from './partnership_icon_gray';
-export * from './portal_icon';
-export * from './portal_icon_gray';
-export * from './project_usage_icon';
-export * from './project_usage_icon_gray';
-export * from './product_activation_icon_gray';
-export * from './team_members_icon';
-export * from './team_members_icon_gray';
