+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/features/project/containers/SideMenu/utils/navigationIcons.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import * as NavigationMenuIcons from '~/assets/navigation-menu';
+import * as NavigationMenuIcons from '~/assets/NavigationMenu';
 
 const navigationIcons = {
 	analytics: [
