+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/features/project/pages/Project/ProjectUsage/components/AddOnContent/AddOnContent.tsx
@@ -4,7 +4,7 @@
  */
 
 import React from 'react';
-import PlusSquaresIcons from '~/assets/plus-squares-icon';
+import {PlusSquaresIcons} from '~/assets/PlusSquaresIcon';
 
 import './AddOnContent.css';
